Tank Shooting
Made with Unity. If we talk about the game in general, we select our tank from the login screen and enter the game. As we hit targets with our tank, the number of enemy tanks decreases and we gain gold. With the gold we earn, we can buy improvements to the tank that I have not designed yet. Our tank is getting better. If we press the P key during the game, our game stops completely. There are some situations that I haven't added to the game yet due to lack of time. One of these is that a completely random life object appears in certain places in the game at certain seconds and comes to life when touched. Another feature I would add is gaining a shield that blocks the first damage.